search

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -38,7 +39,7 @@ class PostController extends Controller
             'content' => "required"
         ]);
 
-        Post::create($validated);
+        auth()->user()->posts()->create($validated);
 
         return redirect()->route('posts.index');
     }
@@ -56,6 +57,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        Gate::authorize('update', $post);
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -64,12 +66,15 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('update', $post);
         $validated = $request->validate([
             'title' => 'required|max:255',
             'content' => "required"
         ]);
 
         $post->update($validated);
+                
+        
         return redirect()->route('posts.index');
     }
 
@@ -78,13 +83,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('delete', $post);
         $post->delete();
         return to_route('posts.index');
-    }
-    
-    public function search(Request $request)
-    {
-        return view('posts.search');
     }
 }
